refactor: Change startchallenge structure #20
