+++ src/main/java/com/doci/webPrj/user/controller/GroupChallengeController.java
@@ -1,0 +1,5 @@
+package com.doci.webPrj.user.controller;
+
+public class GroupChallengeController {
+    
+}
