+++ src/main/java/com/doci/webPrj/user/service/FreeChallengeServiceImp.java
@@ -2,30 +2,19 @@ package com.doci.webPrj.user.service;
 
 import java.time.LocalDate;
 import java.time.Period;
-import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.doci.webPrj.admin.entity.RandomChallenge;
-import com.doci.webPrj.common.repository.RandomChallengeRepository;
-import com.doci.webPrj.user.entity.Choice;
 import com.doci.webPrj.user.entity.FreeChallenge;
-import com.doci.webPrj.user.repository.ChoiceRepository;
 import com.doci.webPrj.user.repository.FreeChallengeRepository;
-import com.doci.webPrj.user.repository.InvitationRepository;
 
 @Service
-public class StartChallengeServiceImp implements StartChallengeService {
-
+public class FreeChallengeServiceImp implements FreeChallengeService {
+    
     @Autowired
     FreeChallengeRepository freeChallengeRepository;
-    @Autowired
-    RandomChallengeRepository randomChallengeRepository;
-    @Autowired
-    ChoiceRepository choiceRepository;
-    
-
+    @Override
     public void addFreeChallenge(FreeChallenge freeChallenge) {
         LocalDate startDate = freeChallenge.getStartDate();
         LocalDate endDate = freeChallenge.getEndDate();
@@ -38,16 +27,6 @@ public class StartChallengeServiceImp implements StartChallengeService {
     }
 
     @Override
-    public void addRandomChallenge(Choice choice) {
-        choiceRepository.save(choice);
-    }
-
-    @Override
-    public List<RandomChallenge> getRandomList(String[] categoryIdList) {
-        return randomChallengeRepository.findRandomByCategory(categoryIdList);
-    }
-
-    @Override
     public int getFreechallengeId(String challengeName,int userId) {
         // String challengeName
        return freeChallengeRepository.getChallengeId(challengeName,userId);
@@ -57,6 +36,6 @@ public class StartChallengeServiceImp implements StartChallengeService {
     public void delete(int challengeId) {
         freeChallengeRepository.delete(challengeId);
     }
-
     
+
 }
