+++ src/main/resources/templates/user/startchallenge/choice/randomcategory.html
@@ -33,7 +33,7 @@
     </div>
     <form
       id="check-form"
-      th:action="@{/challenge/start/choice/randomchallenge}"
+      th:action="@{/randomChallenge/choice/randomchallenge}"
       action=""
       method="post"
     >
