+++ src/main/resources/templates/user/startchallenge/choice/randomchallenge.html
@@ -15,7 +15,7 @@
   <body>
     <header>
       <div class="header-left">
-        <button onclick="location.href='/challenge/start/choice/randomcategory'">
+        <button onclick="location.href='/randomChallenge/choice/randomcategory'">
           <img src="/image/header/blackBack.png" alt="뒤로가기" />
         </button>
         <!-- <img src="/image/title.png" alt="4℃" width="60" height="40" /> -->
@@ -29,7 +29,7 @@
       </h1>
     </section>
 
-    <form class="random-list" method="post" action="/challenge/start/randomchallenge/reg">
+    <form class="random-list" method="post" action="/randomChallenge/randomchallenge/reg">
       <!--    challenge-option 바로 위에 있었는데 옮겼음-->
       <section class="start-order2">
         <div class="date-input">
