+++ src/main/resources/templates/user/startchallenge/group-invite.html
@@ -16,7 +16,7 @@
           <img src="image/header/blackBack.png" alt="뒤로가기" />
         </button>
       </header>
-      <form action="/challenge/start/group-invite/reg" method="post">
+      <form action="/freeChallenge/group-invite/reg" method="post">
         <!-- <div class="group_title">
           <label for="group_name">그룹명</label>
           <div>
